fix(psalm): drop redundant int casts on the request byte deltas
max(0, int - int) is already int<0,max>, so the (int) casts on $bytesIn/$bytesOut
tripped Psalm's RedundantCast (PHPStan does not flag it, so this only surfaced in
CI). Behavior-identical.

+++ src/Application.php
@@ -1386,8 +1386,8 @@ final class Application
                 // Record the request on EVERY path (success, static file, early
                 // return, or exception), mirroring the server's HttpHandler.
                 if ($collector !== null) {
-                    $bytesIn = (int) max(0, $connection->bytesRead - $startBytesRead);
-                    $bytesOut = (int) max(0, $connection->bytesWritten - $startBytesWritten);
+                    $bytesIn = max(0, $connection->bytesRead - $startBytesRead);
+                    $bytesOut = max(0, $connection->bytesWritten - $startBytesWritten);
                     $elapsedMs = (microtime(true) - $startTime) * 1000;
                     $collector->recordRequest(
                         $request->method(),
